Added support for symfony 1.x lib in shared directory.
I prefer to link to symfony lib instead of keeping it in the repository.
This commit introduces two new tasks:
* shared:symfony:download downloads symfony framework and stores it in shared
* symfony:create_lib_symlink creates a lib/vendor/symfony symlink to sources in shared

Download is called after deploy:setup to ensure that symfony is
accessible during the deploy.

Symlink is created before finalizing the update as it is needed
for every symfony task.

No task is automatically run unless use_shared_symfony is set to true.
By default it's false.

As there is no symfony in the repository we cannot guess the version and
have to ask it. Default is provided.