Enhance deploy:set_permissions to grant permissions to the shell user
If the webserver creates new directories in, for example, the cache
directory, the shell user won't have permission to delete those files.
This causes deploy:cleanup to fail, unless sudo is available.

To solve this, the deploy:set_permissions task also has to set ACEs to
ensure that the shell user has full access to the "writable_dirs", just
like it does for the webserver user. This is also what the Symfony 2
documentation says.

As the "chmod +a" command used on OS X can only handle one ACE per
invocation, we have to use an array to support more than one command per
method.